refactor: ConfigValidator tested in own file
